+++ attributes/default.rb
@@ -117,6 +117,13 @@ default['openstack']['image']['api']['swift']['enable_snet'] = 'False'
 default['openstack']['image']['api']['swift']['store_region'] = nil
 default['openstack']['image']['api']['cache']['image_cache_max_size'] = '10737418240'
 
+# Info to match when looking for cinder in the service catalog
+default['openstack']['image']['api']['block-storage']['cinder_catalog_info'] = 'volume:cinder:publicURL'
+# Allow to perform insecure SSL requests to cinder (boolean value)
+default['openstack']['image']['api']['block-storage']['cinder_api_insecure'] = false
+# Location of ca certicates file to use for cinder client requests
+default['openstack']['image']['api']['block-storage']['cinder_ca_certificates_file'] = nil
+
 # Directory for the Image Cache
 default['openstack']['image']['cache']['dir'] = '/var/lib/glance/image-cache/'
 # Number of seconds until an incomplete image is considered stalled an